Rename details view folder
