+++ src/modules/details/views/components/DetailsCard.tsx
